new version

+++ src/tbitfield.cpp
@@ -43,7 +43,7 @@ int TBitField::GetMemIndex(const int n) const // индекс Мем для би
 TELEM TBitField::GetMemMask(const int n) const // битовая маска для бита n
 {
     int bitsInElem = sizeof(TELEM) * 8;
-    return 1 << (n % bitsInElem);
+    return static_cast<unsigned __int64>(1) << (n % bitsInElem);
 }
 
 // доступ к битам битового поля
